refactor: Update User model to include authToken field
This commit adds the `authToken` field to the `User` model in the Prisma schema. The `authToken` field is nullable and will be used to store the authentication token for the user. This change allows for the storage and retrieval of the authentication token in the database, enabling features such as session management and secure authentication.