refactor: Add imageUrl field to LearningResources model
This commit adds the imageUrl field to the LearningResources model in the prisma/schema.prisma file. The imageUrl field is of type String and is nullable. This change enhances the functionality of the application by allowing learning resources to have an associated image URL, which can be used for displaying images in the UI.